Refactor code structure for improved readability and maintainability

+++ FIUBench/evaluate_util.py
@@ -164,9 +164,9 @@ def eval_perturbation_ratio(cfg, tokenizer, eval_dataloader, perturb_dataloader,
  
     
         # zip index and each stat into a dict
-        perturb_loss_per_token = dict(zip(indices, perturb_loss_per_token.cpu().numpy().tolist()))
-        gt_loss_per_token = dict(zip(indices, gt_loss_per_token.cpu().numpy().tolist()))
-        truth_ratio = dict(zip(indices, truth_ratio.cpu().numpy().tolist()))
+        perturb_loss_per_token = dict(zip(indices, perturb_loss_per_token.cpu().float().numpy().tolist()))
+        gt_loss_per_token = dict(zip(indices, gt_loss_per_token.cpu().float().numpy().tolist()))
+        truth_ratio = dict(zip(indices, truth_ratio.cpu().float().numpy().tolist()))
         gt_loss = dict(zip(indices, gt_loss.cpu().numpy().tolist()))
         perturb_loss = dict(zip(indices, perturb_loss.cpu().numpy().tolist()))
         num_token_gt = dict(zip(indices, num_token_gt.cpu().numpy().tolist()))
